Edited, but Done

+++ Soal-[2].py
@@ -1,5 +1,5 @@
 phi = 22/7
-r = int(input("Masukkan jari-jari lingkaran: "))
+r = float(input("Masukkan jari-jari lingkaran: "))
 L = phi*r**2
 
 print("Luas lingkaran dengan jari-jari {}cm adalah {}cm\u00b2".format(r, L))
